Rename `tooltipClasses` to `chartsTooltipClasses`
