+++ scripts/x-charts.exports.json
@@ -48,6 +48,7 @@
   { "name": "ChartsReferenceLineClasses", "kind": "Interface" },
   { "name": "ChartsReferenceLineClassKey", "kind": "TypeAlias" },
   { "name": "ChartsTooltip", "kind": "Function" },
+  { "name": "chartsTooltipClasses", "kind": "Variable" },
   { "name": "ChartsTooltipClasses", "kind": "Interface" },
   { "name": "ChartsTooltipClassKey", "kind": "TypeAlias" },
   { "name": "ChartsTooltipProps", "kind": "TypeAlias" },
@@ -77,6 +78,7 @@
   { "name": "getAreaElementUtilityClass", "kind": "Function" },
   { "name": "getAxisHighlightUtilityClass", "kind": "Function" },
   { "name": "getAxisUtilityClass", "kind": "Function" },
+  { "name": "getChartsTooltipUtilityClass", "kind": "Function" },
   { "name": "getHighlightElementUtilityClass", "kind": "Function" },
   { "name": "getLegendUtilityClass", "kind": "Function" },
   { "name": "getLineElementUtilityClass", "kind": "Function" },
@@ -84,7 +86,6 @@
   { "name": "getPieArcLabelUtilityClass", "kind": "Function" },
   { "name": "getPieArcUtilityClass", "kind": "Function" },
   { "name": "getReferenceLineUtilityClass", "kind": "Function" },
-  { "name": "getTooltipUtilityClass", "kind": "Function" },
   { "name": "getValueToPositionMapper", "kind": "Function" },
   { "name": "HighlightElementClassKey", "kind": "TypeAlias" },
   { "name": "HighlightOptions", "kind": "TypeAlias" },
@@ -145,7 +146,6 @@
   { "name": "StackableSeriesType", "kind": "TypeAlias" },
   { "name": "StackOffsetType", "kind": "TypeAlias" },
   { "name": "StackOrderType", "kind": "TypeAlias" },
-  { "name": "tooltipClasses", "kind": "Variable" },
   { "name": "useDrawingArea", "kind": "Function" },
   { "name": "useXScale", "kind": "Function" },
   { "name": "useYScale", "kind": "Function" }
